【********】域维度层次化分析性能差异-生成Megetron布局2
# message auto-generated for no-merge-commit merge:
merge master_parallelism_megatron_1219 into master

【********】域维度层次化分析性能差异-生成Megetron布局2

Created-by: insightDeveloper

Match-id-82f6afee014333714f039af9edc5f583778b338e

+++ server/src/modules/communication/parser/ClusterDef.h
@@ -182,8 +182,8 @@ const std::vector<std::string> PARALLEL_STRATEGY_HEADERS = {
 };
 
 struct Position {
-    uint8_t x = 0;
-    uint8_t y = 0;
+    uint32_t x = 0;
+    uint32_t y = 0;
     bool operator==(const Position& other) const
     {
         return x == other.x && y == other.y;
